fix(upgrade): Move upgrade info to a dependency on CodeSystems
