+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/codesystem/CodeSystem.java
@@ -51,6 +51,10 @@ public final class CodeSystem extends TerminologyResource {
 		return codeSystemIdWithQuery.startsWith(RESOURCE_TYPE + Branch.SEPARATOR) ? new ResourceURIWithQuery(codeSystemIdWithQuery) : ResourceURIWithQuery.of(RESOURCE_TYPE, codeSystemIdWithQuery);
 	}
 	
+	public UpgradeInfo getUpgradeInfo() {
+		return getDependency(TerminologyResource.DependencyScope.UPGRADE_OF).map(Dependency::getUpgradeInfo).orElse(null);
+	}
+	
 	public static CodeSystem from(ResourceDocument doc) {
 		CodeSystem codeSystem = new CodeSystem();
 		codeSystem.setId(doc.getId());
